[nrf fromlist] linker: arm cortex linker script _flash_used fix
linker.ld contains a `.last_section` section that is placed in rom
region as NOLOAD for the purpose of retrieve the actual number of bytes
contained in the image. See d85efe0b10f1f8ed4b854717f994062a4bbffe26

However, a previous section may cause the location counter to be
incremented for alignment purposes. This can result in the size of the
image to be 0x10FA but location counter to be 0x1100 because it has been
aligned for next section placement.

By removing NOLOAD and write 0xE015 (end of last section) to the section
we ensure that data is written after alignment of location counter, and
thereby forces the image size to be in sync with the location counter.

Using 0xE015 also serves as identifier that end of last section is
present.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>