[nrf toup] sysbuild: support specifying remote board through Kconfig
Introducing `<application-setting>_BOARD` pattern which can be used for
additional image support to adjust the remote board.

Having a Kconfig following a pattern as `<application-setting>_BOARD`
allows board Kconfig or out-of-tree Kconfig trees to adjust the remote
board when specific Kconfig settings are met.

Such as use-case can be to set the remote board when a specific board
is used for the main application.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>