[nrf toup] sysbuild: support specifying hci rpmsg name through Kconfig
Introducing `<application-setting>_NAME` pattern which can be used for
additional image support to adjust the image name as well as be able
to fetch the image name of a given image.

Having a Kconfig following a pattern as `<application-setting>_NAME`
allows board Kconfig or out-of-tree Kconfig trees to adjust the name
if needed, for example to avoid name collisions.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>